remove remnants of old sync
