+++ sync/src/main/kotlin/gropius/sync/model/LabelInfo.kt
@@ -1,25 +1,24 @@
-package gropius.sync.github.model
+package gropius.sync.model
 
 import org.bson.types.ObjectId
 import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.index.Indexed
 import org.springframework.data.mongodb.core.mapping.Document
-import java.net.URI
 
 /**
- * Mapping of a single user from neo4j to GitHub
+ * Mapping of a single label from neo4j to GitHub
  * @param url API URL of IMS of the repo
- * @param login username on GitHub
- * @param neo4jId IMSUser ID in gropius database
+ * @param githubId ID on GitHub
+ * @param neo4jId ID in gropius database
  */
 @Document
-data class UserInfo(
+data class LabelInfo(
     @Indexed
-    val login: String,
+    val imsProject: String,
+    @Indexed
+    val githubId: String,
     @Indexed(unique = true)
-    val neo4jId: String,
-    @Indexed
-    val url: URI
+    val neo4jId: String
 ) {
     /**
      * MongoDB ID
